Collate all cart utils into cart provider
