+++ providers/CartProvider.js
@@ -16,14 +16,65 @@ const getCartItemsFromLocalStorage = () => {
   return cartItems ? JSON.parse(cartItems) : [];
 };
 
+const appendToLocalStorageCart = (currentLocalStorageCart, cartItem) => {
+  // find and merge duplicate items
+  let hasDuplicates = false;
+  const newLocalStorageCart = currentLocalStorageCart.map((item) => {
+    if (item.id === cartItem.id && item.expiryId === cartItem.expiryId) {
+      hasDuplicates = true;
+      return {
+        ...item,
+        cartQuantity: item.cartQuantity + cartItem.cartQuantity,
+      };
+    }
+    return item;
+  });
+  // if no duplicates, append to cart
+  if (!hasDuplicates) {
+    newLocalStorageCart.push(cartItem);
+  }
+  return newLocalStorageCart;
+};
+
 export const CartProvider = ({ children }) => {
   const [cartState, setCartState] = useState(getCartStateFromLocalStorage());
   const [cartItems, setCartItems] = useState(
     getCartItemsFromLocalStorage() || [],
   );
+
+  const addToCart = (cartItem) => {
+    const localStorageCartState = appendToLocalStorageCart(cartItems, cartItem);
+    localStorage.setItem(
+      LOCAL_STORAGE_CART_KEY,
+      JSON.stringify(localStorageCartState),
+    );
+    setCartItems(localStorageCartState);
+    setCartState(localStorageCartState[0].type);
+  };
+
+  const clearCart = () => {
+    localStorage.removeItem(LOCAL_STORAGE_CART_KEY);
+    setCartItems([]);
+    setCartState('');
+  };
+
+  const removeItemFromCart = (expiryId) => {
+    const newCartItems = cartItems.filter((item) => item.expiryId !== expiryId);
+    if (newCartItems.length === 0) {
+      localStorage.removeItem(LOCAL_STORAGE_CART_KEY);
+      setCartItems([]);
+      setCartState('');
+      return;
+    }
+
+    localStorage.setItem(LOCAL_STORAGE_CART_KEY, JSON.stringify(newCartItems));
+    setCartItems(newCartItems);
+    setCartState(newCartItems[0].type);
+  };
+
   return (
     <CartContext.Provider
-      value={{ cartState, setCartState, cartItems, setCartItems }}
+      value={{ cartState, cartItems, addToCart, clearCart, removeItemFromCart }}
     >
       {children}
     </CartContext.Provider>
