--- utils/cart-utils/clearCart.js
@@ -1,6 +1,0 @@
-import { LOCAL_STORAGE_CART_KEY } from '../../globals';
-
-export const clearCart = (setCartState) => {
-  localStorage.removeItem(LOCAL_STORAGE_CART_KEY);
-  setCartState();
-};
