--- utils/cart-utils/index.js
@@ -1,4 +1,0 @@
-export * from './addToCart';
-export * from './clearCart';
-export * from './getCartState';
-export * from './removeItemFromCart';
