+++ utils/index.js
@@ -1,4 +1,3 @@
-export * from './cart-utils';
 export * from './getDjangoFriendlyDate';
 export * from './levenshteinDistance';
 export * from './submitForm';
